move files around and adjust components
