+++ components/footer/Footer.tsx
@@ -1,6 +1,6 @@
-import Icon from "./ui/Icon.tsx";
+import Icon from "$store/components/ui/Icon.tsx";
 
-export const sections = [
+const sections = [
   {
     label: "Sobre",
     content: (
@@ -125,21 +125,84 @@ export const sections = [
   },
 ];
 
-function FooterAccordion() {
+function Footer() {
   return (
-    <div class="w-full py-3 bg-[#353535]">
-      {sections.map(({ label, content }) => (
-        <div key={label}>
-          <details class="w-full cursor-pointer flex flex-row justify-between focus:outline-none">
-            <summary class="text-white px-6 py-2">{label}</summary>
-            <div class="px-8 text-sm text-gray-100">
+    <footer class="w-full">
+      <div class="hidden md:block bg-[#353535] w-full">
+        <div class="flex justify-between p-8 w-9/12 mx-auto">
+          {sections.map(({ label, content }) => (
+            <div class="flex flex-col mr-12">
+              <span class="text-white uppercase mb-4 text-xs">{label}</span>
               {content}
             </div>
-          </details>
+          ))}
         </div>
-      ))}
-    </div>
+      </div>
+
+      <div class="md:hidden w-full">
+        <div class="w-full py-3 bg-[#353535]">
+          {sections.map(({ label, content }) => (
+            <div key={label}>
+              <details class="w-full cursor-pointer flex flex-row justify-between focus:outline-none">
+                <summary class="text-white px-6 py-2">{label}</summary>
+                <div class="px-8 text-sm text-gray-100">
+                  {content}
+                </div>
+              </details>
+            </div>
+          ))}
+        </div>
+      </div>
+
+      <div className="flex flex-row flex-col justify-between justify-center items-center px-3 pt-2 pb-5 w-full">
+        <div class="flex flex-col items-center">
+          <ul class="flex items-center justify-center mt-3">
+            <li class="mr-[8px] last:m-0">
+              <a
+                href="#"
+                class="flex items-center justify-center w-8 h-8"
+                target="_blank"
+                rel="noopener noreferrer"
+                aria-label="Facebook logo"
+              >
+                <Icon id="Facebook" class="w-auto" />
+              </a>
+            </li>
+            <li class="mr-[8px] last:m-0">
+              <a
+                href="https://www.instagram.com/deco.cx"
+                class="flex items-center justify-center w-8 h-8"
+                target="_blank"
+                rel="noopener noreferrer"
+                aria-label="Instagram logo"
+              >
+                <Icon id="Instagram" class="w-auto" />
+              </a>
+            </li>
+            <li class="mr-[8px] last:m-0">
+              <a
+                href="#"
+                class="flex items-center justify-center w-8 h-8"
+                target="_blank"
+                rel="noopener noreferrer"
+                aria-label="Tiktok logo"
+              >
+                <Icon id="Tiktok" class="w-auto" />
+              </a>
+            </li>
+          </ul>
+        </div>
+      </div>
+
+      <div class="w-full">
+        <div class="flex justify-center px-3 pb-4">
+          <p class="text-xs text-center">
+            Site by <a href="https://www.deco.cx">deco.cx</a>
+          </p>
+        </div>
+      </div>
+    </footer>
   );
 }
 
-export default FooterAccordion;
+export default Footer;
