+++ fresh.gen.ts
@@ -20,10 +20,10 @@ import * as $$6 from "./islands/SearchControls.tsx";
 import * as $$7 from "./islands/Slider.tsx";
 import * as $$$0 from "./sections/Button.story.tsx";
 import * as $$$1 from "./sections/Carousel.tsx";
-import * as $$$2 from "./sections/Footer.tsx";
-import * as $$$3 from "./sections/Head.tsx";
-import * as $$$4 from "./sections/Header.tsx";
-import * as $$$5 from "./sections/HighlightedFeatures.tsx";
+import * as $$$2 from "./sections/Features.tsx";
+import * as $$$3 from "./sections/Footer.tsx";
+import * as $$$4 from "./sections/Head.tsx";
+import * as $$$5 from "./sections/Header.tsx";
 import * as $$$6 from "./sections/Highlights.tsx";
 import * as $$$7 from "./sections/Newsletter.tsx";
 import * as $$$8 from "./sections/ProductDetails.tsx";
@@ -31,8 +31,7 @@ import * as $$$9 from "./sections/ProductGallery.tsx";
 import * as $$$10 from "./sections/ProductShelf.tsx";
 import * as $$$11 from "./sections/SearchControls.tsx";
 import * as $$$12 from "./sections/Spacer.tsx";
-import * as $$$13 from "./sections/StoreFeatures.tsx";
-import * as $$$14 from "./sections/vtexconfig.global.tsx";
+import * as $$$13 from "./sections/vtexconfig.global.tsx";
 import * as $$$$0 from "./functions/occProductDetailsPage.ts";
 import * as $$$$1 from "./functions/shopifyProductDetailsPage.ts";
 import * as $$$$2 from "./functions/shopifyProductList.ts";
@@ -66,10 +65,10 @@ const manifest: DecoManifest = {
   sections: {
     "./sections/Button.story.tsx": $$$0,
     "./sections/Carousel.tsx": $$$1,
-    "./sections/Footer.tsx": $$$2,
-    "./sections/Head.tsx": $$$3,
-    "./sections/Header.tsx": $$$4,
-    "./sections/HighlightedFeatures.tsx": $$$5,
+    "./sections/Features.tsx": $$$2,
+    "./sections/Footer.tsx": $$$3,
+    "./sections/Head.tsx": $$$4,
+    "./sections/Header.tsx": $$$5,
     "./sections/Highlights.tsx": $$$6,
     "./sections/Newsletter.tsx": $$$7,
     "./sections/ProductDetails.tsx": $$$8,
@@ -77,8 +76,7 @@ const manifest: DecoManifest = {
     "./sections/ProductShelf.tsx": $$$10,
     "./sections/SearchControls.tsx": $$$11,
     "./sections/Spacer.tsx": $$$12,
-    "./sections/StoreFeatures.tsx": $$$13,
-    "./sections/vtexconfig.global.tsx": $$$14,
+    "./sections/vtexconfig.global.tsx": $$$13,
   },
   functions: {
     "./functions/occProductDetailsPage.ts": $$$$0,
@@ -215,6 +213,49 @@ const manifest: DecoManifest = {
           },
         },
         "required": [],
+      },
+      "outputSchema": null,
+    },
+    "./sections/Features.tsx": {
+      "inputSchema": {
+        "title": " Features",
+        "type": "object",
+        "properties": {
+          "features": {
+            "type": "array",
+            "items": {
+              "title": "Feature",
+              "type": "object",
+              "properties": {
+                "src": {
+                  "format": "image-uri",
+                  "type": "string",
+                  "title": "Src",
+                  "description": "Image src",
+                },
+                "title": {
+                  "type": "string",
+                  "title": "Title",
+                  "description": "Title",
+                },
+                "description": {
+                  "type": "string",
+                  "title": "Description",
+                  "description": "Description and Image alt text",
+                },
+              },
+              "required": [
+                "src",
+                "title",
+                "description",
+              ],
+            },
+            "title": "Features",
+          },
+        },
+        "required": [
+          "features",
+        ],
       },
       "outputSchema": null,
     },
@@ -388,49 +429,6 @@ const manifest: DecoManifest = {
       },
       "outputSchema": null,
     },
-    "./sections/HighlightedFeatures.tsx": {
-      "inputSchema": {
-        "title": " Highlighted Features",
-        "type": "object",
-        "properties": {
-          "features": {
-            "type": "array",
-            "items": {
-              "title": "Feature",
-              "type": "object",
-              "properties": {
-                "src": {
-                  "format": "image-uri",
-                  "type": "string",
-                  "title": "Src",
-                  "description": "Image src",
-                },
-                "title": {
-                  "type": "string",
-                  "title": "Title",
-                  "description": "Title",
-                },
-                "description": {
-                  "type": "string",
-                  "title": "Description",
-                  "description": "Description and Image alt text",
-                },
-              },
-              "required": [
-                "src",
-                "title",
-                "description",
-              ],
-            },
-            "title": "Features",
-          },
-        },
-        "required": [
-          "features",
-        ],
-      },
-      "outputSchema": null,
-    },
     "./sections/Highlights.tsx": {
       "inputSchema": {
         "title": " Highlights",
@@ -587,10 +585,6 @@ const manifest: DecoManifest = {
         },
         "required": [],
       },
-      "outputSchema": null,
-    },
-    "./sections/StoreFeatures.tsx": {
-      "inputSchema": null,
       "outputSchema": null,
     },
     "./sections/vtexconfig.global.tsx": {
