+++ islands/LiveControls.tsx
@@ -1,3 +1,1 @@
-import LiveControls from "$live/components/LiveControls.tsx";
-
-export default LiveControls;
+export { default } from "$live/components/LiveControls.tsx";
