+++ islands/MenuModal.tsx
@@ -1,4 +1,1 @@
-import MenuModal from "../components/menu/MenuModal.tsx";
-import type { NavItem } from "../components/menu/NavItem.ts";
-
-export default ({ items }: { items: NavItem[] }) => <MenuModal items={items} />;
+export { default } from "$store/components/header/MenuModal.tsx";
