+++ sections/Features.tsx
@@ -1,0 +1,3 @@
+export { default } from "$store/components/ui/Features.tsx";
+export type { Props, Feature } from "$store/components/ui/Features.tsx";
+export type { Image as LiveImage } from "$live/std/ui/types/Image.ts";
