+++ sections/SearchControls.tsx
@@ -1,18 +1,8 @@
-import { ProductListingPage } from "$live/std/commerce/types.ts";
-import { LoaderReturnType } from "$live/std/types.ts";
+import SearchControls from "$store/islands/SearchControls.tsx";
+import type { Props } from "$store/components/search/Controls.tsx";
 
-import SearchControlsIsland from "../islands/SearchControls.tsx";
-
-export interface Props {
-  page: LoaderReturnType<ProductListingPage>;
+function SearchControlsSection(props: Props) {
+  return <SearchControls {...props} />;
 }
 
-export default function SearchControls({ page }: Props) {
-  const filters = page?.filters;
-
-  if (!filters || filters.length === 0) {
-    return null;
-  }
-
-  return <SearchControlsIsland filters={filters} />;
-}
+export default SearchControlsSection;
