+++ sections/Spacer.tsx
@@ -1,15 +1,2 @@
-export interface Props {
-  height?: {
-    mobile: number;
-    desktop: number;
-  };
-}
-
-function Spacer({ height }: Props) {
-  const mobile = height?.mobile ?? 8;
-  const desktop = height?.desktop ?? 8;
-
-  return <div class={`h-${mobile} md:h-${desktop}`} />;
-}
-
-export default Spacer;
+export { default } from "$store/components/ui/Spacer.tsx";
+export type { Props } from "$store/components/ui/Spacer.tsx";
